+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 Extending the adopted spec, each change should have a link to its
 corresponding pull request appended.
 
-## [1.0.0] - 2021-03-18
+## [1.0.0] - 2022-04-30
 
-This is the initial release of the module, with support for <module_name_here> resources.
+This is the initial release of the module, with support for Cloud Pak for Integration resources.
+
+Can have following 3 sections
+
+## v1.1 (TBD)
+
+FEATURES:
+
+* **new feature:**  Adding support for custom namespace
+
+* **new feature:**  Adding support for LDAP automation
+
+## Sample Changelog
+
+* FEATURES:      Description (pullRequest link)
+* IMPROVEMENTS:  Description (pullRequest link)
+* BUGS:          Description (pullRequest link)